Refactor XmpParser class and MotionPhotoInfo class.
