+++ MotionPhotoReader/app/src/main/java/com/example/motionphotoreader/MotionPhotoInfo.java
@@ -1,0 +1,9 @@
+package com.example.motionphotoreader;
+
+/**
+ * Holds data about the Motion Photo file.
+ */
+public class MotionPhotoInfo {
+
+
+}
