+++ MotionPhotoReader/app/src/main/java/com/example/motionphotoreader/MotionPhotoReader.java
@@ -53,6 +53,7 @@ public class MotionPhotoReader {
     private static final int MSG_SEEK_TO_FRAME = 0x0010;
     private static final long TIMEOUT_MS = 1000L;
 
+    private static final long MS_TO_NS = 1000L;
     /**
      * Two handlers manage the calls to play through the video. The media worker thread posts
      * available buffers to the buffer queue. The buffer worker receives messages to process frames
@@ -199,7 +200,7 @@ public class MotionPhotoReader {
                         }
                             else {
                             Log.d("NextFrame", "Queue InputBuffer for time " + lowResExtractor.getSampleTime());
-                            lowResDecoder.queueInputBuffer(bufferIndex, 0, sampleSize, lowResExtractor.getSampleTime(), 0);
+                            lowResDecoder.queueInputBuffer(bufferIndex, 0, sampleSize, lowResExtractor.getSampleTime() * MS_TO_NS, 0);
                             lowResExtractor.advance();
                         }
 
@@ -211,7 +212,7 @@ public class MotionPhotoReader {
                         }
                         bufferIndex = bufferData.getInt("BUFFER_INDEX");
                         timestamp = bufferData.getLong("TIMESTAMP");
-                        lowResDecoder.releaseOutputBuffer(bufferIndex, timestamp * 1000L);
+                        lowResDecoder.releaseOutputBuffer(bufferIndex, timestamp * MS_TO_NS);
                         Log.d("NextFrame", "Releasing to output buffer " + bufferIndex);
                         break;
 
@@ -238,7 +239,7 @@ public class MotionPhotoReader {
                         }
                         else {
                             Log.d("SeekToFrame", "Queue InputBuffer " + lowResExtractor.getSampleTime());
-                            lowResDecoder.queueInputBuffer(bufferIndex, 0, sampleSize, lowResExtractor.getSampleTime(), 0);
+                            lowResDecoder.queueInputBuffer(bufferIndex, 0, sampleSize, lowResExtractor.getSampleTime() * MS_TO_NS, 0);
                             lowResExtractor.advance();
                         }
 
@@ -250,7 +251,7 @@ public class MotionPhotoReader {
                         }
                         bufferIndex = bufferData.getInt("BUFFER_INDEX");
                         timestamp = bufferData.getLong("TIMESTAMP");
-                        lowResDecoder.releaseOutputBuffer(bufferIndex, timestamp * 1000L);
+                        lowResDecoder.releaseOutputBuffer(bufferIndex, timestamp * MS_TO_NS);
                         Log.d("SeekToFrame", "Releasing to output buffer " + bufferIndex);
                         break;
 
@@ -344,84 +345,6 @@ public class MotionPhotoReader {
         message.setData(messageData);
 
         message.sendToTarget();
-    }
-
-    /**
-     * The XmpParser class is a child class intended to help extract the microvideo offset information
-     * from the XMP metadata of the given Motion Photo. This class is for internal use only.
-     */
-    private static class XmpParser {
-
-        private static final byte[] OPEN_ARR = "<x:xmpmeta".getBytes();  /* Start of XMP metadata tag */
-        private static final byte[] CLOSE_ARR = "</x:xmpmeta>".getBytes();  /* End of XMP metadata tag */
-
-        /**
-         * Copies the input stream from a file to an output stream.
-         */
-        private static void copy(String filename, InputStream in, OutputStream out) throws IOException {
-            int len;
-            byte[] buf = new byte[1024];
-            while((len = in.read(buf)) >= 0) {
-                out.write(buf, 0, len);
-            }
-            in.close();
-            out.close();
-        }
-
-        /**
-         * Returns the index of the first appearance of a given subsequence in a byte array.
-         * @param arr The full byte array
-         * @param seq The subsequence of bytes to find
-         * @param start The index at which to start searching
-         * @return The index of the first appearance of the beginning of the subsequence in the
-         * byte array. If the sequence is not found, return -1.
-         */
-        private static int indexOf(byte[] arr, byte[] seq, int start) {
-            int subIdx = 0;
-            for (int x = start; x < arr.length; x++) {
-                if (arr[x] == seq[subIdx]) {
-                    if (subIdx == seq.length - 1) {
-                        return x - subIdx;
-                    }
-                    subIdx++;
-                }
-                else {
-                    subIdx = 0;
-                }
-            }
-            return -1;
-        }
-
-        /**
-         * Returns the video offset of the microvideo in the Motion Photo file, in bytes from the end
-         * of the file.
-         */
-        public static int getVideoOffset(String filename) throws XMPException, IOException {
-            FileInputStream in = new FileInputStream(filename);
-            ByteArrayOutputStream out = new ByteArrayOutputStream();
-
-            copy(filename, in, out);
-            byte[] fileData = out.toByteArray();
-
-            int openIdx = indexOf(fileData, OPEN_ARR, 0);
-            if (openIdx >= 0) {
-                int closeIdx = indexOf(fileData, CLOSE_ARR, openIdx + 1) + CLOSE_ARR.length;
-
-                byte[] segArr = Arrays.copyOfRange(fileData, openIdx, closeIdx);
-                XMPMeta meta = XMPMetaFactory.parseFromBuffer(segArr);
-
-                int videoOffset = meta.getPropertyInteger("http://ns.google.com/photos/1.0/camera/", "MicroVideoOffset");
-                Log.d("XmlParserActivity", "Micro video offset: " + videoOffset);
-                return videoOffset;
-            }
-            return 0;
-
-        }
-
-    }
-    
-    private class MotionPhotoInfo {
-        
     }
     
     private class MotionPhotoImage {
